Fix warnings

+++ samples/dokan_memory/FileNode.cpp
@@ -1,27 +1,17 @@
 #include "FileNode.h"
 
-// Informations
-// FileName can change during move
-
-
-// Data
-// Object lock not needed
-
 FileNode::FileNode(std::wstring fileName, bool isDirectory)
     : FileNode(fileName, isDirectory,
                isDirectory ? FILE_ATTRIBUTE_DIRECTORY : FILE_ATTRIBUTE_ARCHIVE,
                nullptr) {}
 
-FileNode::FileNode(std::wstring fileName, bool isDirectory,
-                          DWORD fileAttr,
-                          PDOKAN_IO_SECURITY_CONTEXT SecurityContext)
+FileNode::FileNode(std::wstring fileName, bool isDirectory, DWORD fileAttr,
+                   PDOKAN_IO_SECURITY_CONTEXT SecurityContext)
     : _fileName(fileName), IsDirectory(isDirectory), Attributes(fileAttr) {
+  // No lock need, FileNode is still not in directory
   Times.Reset();
-  // FilePath = std::filesystem::path(FileName);
 
   if (SecurityContext && SecurityContext->AccessState.SecurityDescriptor) {
-    // Set security information
-    // No lock need, FileNode is still not in directory
     Security.DescriptorSize = GetSecurityDescriptorLength(
         SecurityContext->AccessState.SecurityDescriptor);
     Security.Descriptor = new byte[Security.DescriptorSize];
@@ -35,7 +25,7 @@ DWORD FileNode::Read(LPVOID Buffer, DWORD BufferLength, LONGLONG Offset) {
   if (static_cast<size_t>(Offset + BufferLength) > _data.size())
     BufferLength = static_cast<DWORD>(_data.size() - Offset);
   if (BufferLength)
-    memcpy(Buffer, &_data[Offset], BufferLength);
+    memcpy(Buffer, &_data[static_cast<size_t>(Offset)], BufferLength);
   return BufferLength;
 }
 
@@ -45,19 +35,19 @@ DWORD FileNode::Write(LPCVOID Buffer, DWORD NumberOfBytesToWrite,
 
   std::lock_guard<std::mutex> lock(_data_mutex);
   if (static_cast<size_t>(Offset + NumberOfBytesToWrite) > _data.size())
-    _data.resize(Offset + NumberOfBytesToWrite);
-  memcpy(&_data[Offset], Buffer, NumberOfBytesToWrite);
+    _data.resize(static_cast<size_t>(Offset + NumberOfBytesToWrite));
+  memcpy(&_data[static_cast<size_t>(Offset)], Buffer, NumberOfBytesToWrite);
   return NumberOfBytesToWrite;
 }
 
-const size_t FileNode::getFileSize() {
+const LONGLONG FileNode::getFileSize() {
   std::lock_guard<std::mutex> lock(_data_mutex);
-  return _data.size();
+  return static_cast<LONGLONG>(_data.size());
 }
 
 void FileNode::setEndOfFile(const LONGLONG& byteOffset) {
   std::lock_guard<std::mutex> lock(_data_mutex);
-  _data.resize(byteOffset);
+  _data.resize(static_cast<size_t>(byteOffset));
 }
 
 const std::wstring FileNode::getFileName() {
